refactor: move writeEnvironmentImportsManifest to import-environment.ts
